Made adjustment to chart props include customizability of axis and series encoding configurations
