+++ app/web_templates/src/components/charts/bar.tsx
@@ -2,20 +2,13 @@ import useChart from '@/lib/hooks/chart.ts';
 import ToolBox from '@/components/charts/toolbox.tsx';
 import type { ReactElement, FC } from 'react';
 import type { ChartProps } from '@/components/charts/props.ts';
-import type { XAXisComponentOption, YAXisComponentOption, TooltipComponentOption } from "echarts/types/dist/echarts";
-import type { SeriesOption } from "echarts";
+import type { XAXisComponentOption, YAXisComponentOption, TooltipComponentOption } from 'echarts/types/dist/echarts';
+import type { SeriesOption } from 'echarts';
 import type { Dimensions } from '@/lib/hooks/chart.ts';
 
-export type HistogramBin = {
-    binStart: number;
-    binEnd: number;
-    count: number;
-    label: string;
-};
-
-export default function BarChart({ data, labels, tooltipFormatter }: ChartProps): ReactElement<FC> {
+export default function BarChart({ data, labels, tooltipFormatter, axis, seriesEncoding }: ChartProps): ReactElement<FC> {
     const dimensions: Dimensions = {width: 800, height: 600};
-    const xAxis: XAXisComponentOption = {
+    let xAxis: XAXisComponentOption = {
         type: 'category',
         name: labels != null && labels.x.length > 0 ? labels.x : undefined,
         nameTextStyle: {
@@ -25,7 +18,7 @@ export default function BarChart({ data, labels, tooltipFormatter }: ChartProps)
         axisTick: { show: false },
         splitLine: { show: false },
     };
-    const yAxis: YAXisComponentOption = {
+    let yAxis: YAXisComponentOption = {
         type: 'value',
         name: labels != null && labels.y.length > 0 ? labels.y : undefined,
         nameTextStyle: {
@@ -37,20 +30,18 @@ export default function BarChart({ data, labels, tooltipFormatter }: ChartProps)
         axisLabel: { show: false },
         splitLine: {lineStyle: { color: '#ccc' }}
     };
-    const series: Array<SeriesOption> = [
-        {
-            type: 'bar',
-            datasetIndex: 0,
-            barWidth: '100%',
-            itemStyle: {
-                borderRadius: 2
-            },
-            encode: {
-                x: 'label',
-                y: 'count'
-            }
+    let series: SeriesOption = {
+        type: 'bar',
+        datasetIndex: 0,
+        barWidth: '100%',
+        itemStyle: {
+            borderRadius: 2
+        },
+        encode: {
+            x: 'x',
+            y: 'y'
         }
-    ];
+    };
     const tooltipOptions: TooltipComponentOption = {
         formatter: tooltipFormatter,
         trigger: 'axis',
@@ -58,6 +49,10 @@ export default function BarChart({ data, labels, tooltipFormatter }: ChartProps)
             type: 'shadow',
         }
     };
+
+    if(axis && axis.x) xAxis = {...xAxis, ...axis.x};
+    if(axis && axis.y) yAxis = {...yAxis, ...axis.y};
+    if(seriesEncoding) series.encode = {...series.encode, ...seriesEncoding};
 
     const { containerRef, chart } = useChart(data, xAxis, yAxis, series, tooltipOptions);
