+++ app/web_templates/src/components/charts/props.ts
@@ -1,5 +1,6 @@
 import type { Presenter } from '@/lib/data/presenters';
-import type {DatasetOption, TopLevelFormatterParams} from "echarts/types/dist/shared";
+import type { DatasetOption, TopLevelFormatterParams } from "echarts/types/dist/shared";
+import type { XAXisComponentOption, YAXisComponentOption } from "echarts/types/dist/echarts";
 
 interface ChartContainerProps {
     presenter: Presenter;
@@ -8,6 +9,14 @@ interface ChartContainerProps {
 interface ChartProps {
     data: DatasetOption | Array<DatasetOption>;
     tooltipFormatter?: (params: TopLevelFormatterParams) => string;
+    axis?: {
+        x?: XAXisComponentOption;
+        y?: YAXisComponentOption;
+    };
+    seriesEncoding?: {
+        x?: string | number;
+        y?: string | number;
+    };
     labels?: {
         x: string;
         y: string;
